Rat AI updates. Documentation and comment clarity.
Comments within blueprints updated for clarity.

If RatSwarm is activated, a boolean that checks whether or not the rat was player-spawned is set to true. If it is, they will automatically seek out the nearest AI if Swarm() is eligible.

Additionally, if the rat no longer overlaps with another rat's sphere, a check will be performed on whether or not Swarm() is still eligible.